Move pages to components
Next.js ONLY wants pages in pages/*, and breaks when trying to build
with test and snapshot files.
