+++ components/containerView/index.tsx
@@ -11,9 +11,9 @@ import {
   TableRow,
 } from "@material-ui/core";
 
-import UserContext from "../../src/contexts/UserContext";
 import { ILoggedInSolidSession } from "@inrupt/solid-auth-fetcher/dist/solidSession/ISolidSession";
 import { fetch } from "@inrupt/solid-auth-fetcher";
+import UserContext from "../../src/contexts/UserContext";
 import { useRedirectIfLoggedOut } from "../../src/effects/auth";
 import Spinner from "../spinner";
