+++ components/pages/index/index.test.tsx
@@ -1,21 +1,19 @@
 import { shallow } from "enzyme";
 import { shallowToJson } from "enzyme-to-json";
 
-import { useRedirectIfLoggedOut } from "../../src/effects/auth";
-import LoginSuccessPage from "./index";
+import { useRedirectIfLoggedOut } from "../../../src/effects/auth";
+import IndexPage from "./index";
 
-jest.mock("../../src/effects/auth");
+jest.mock("../../../src/effects/auth");
 
-describe("LoginSuccess page", () => {
-  // TODO test effect
-
+describe("Index page", () => {
   test("Renders a logout button", () => {
-    const tree = shallow(<LoginSuccessPage />);
+    const tree = shallow(<IndexPage />);
     expect(shallowToJson(tree)).toMatchSnapshot();
   });
 
   test("Redirects if the user is logged out", () => {
-    shallow(<LoginSuccessPage />);
+    shallow(<IndexPage />);
     expect(useRedirectIfLoggedOut).toHaveBeenCalled();
   });
 });
