+++ components/pages/login/index.tsx
@@ -1,0 +1,9 @@
+import { ReactElement } from "react";
+import { useRedirectIfLoggedIn } from "../../../src/effects/auth";
+import LoginForm from "../../login";
+
+export default function Login(): ReactElement {
+  useRedirectIfLoggedIn();
+
+  return <LoginForm />;
+}
