+++ constants/__mocks__/next/config/index.ts
