+++ package.json
@@ -23,51 +23,51 @@
   "homepage": "https://github.com/inrupt/pod-manager#readme",
   "dependencies": {
     "@inrupt/solid-auth-fetcher": "0.0.6",
-    "@material-ui/core": "^4.9.14",
+    "@material-ui/core": "^4.10.1",
     "@material-ui/icons": "^4.9.1",
     "encoding": "^0.1.12",
     "eslint-config-prettier": "^6.11.0",
-    "lit-solid": "master",
-    "next": "^9.4.0",
+    "lit-solid": "0.0.2-master-124600945-506.0",
+    "next": "^9.4.4",
     "next-runtime-dotenv": "^1.2.0",
     "prop-types": "^15.7.2",
     "rdf-namespaces": "^1.8.0",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
-    "react-jss": "^10.1.1",
+    "react-jss": "^10.2.0",
     "react-transition-group": "~4.3.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.9.6",
-    "@babel/plugin-proposal-class-properties": "^7.8.3",
-    "@babel/preset-typescript": "^7.9.0",
+    "@babel/core": "^7.10.2",
+    "@babel/plugin-proposal-class-properties": "^7.10.1",
+    "@babel/preset-typescript": "^7.10.1",
     "@types/enzyme": "^3.10.5",
     "@types/enzyme-adapter-react-16": "^1.0.6",
-    "@types/jest": "^25.2.1",
+    "@types/jest": "^25.2.3",
     "@types/react": "^16.9.35",
-    "@typescript-eslint/eslint-plugin": "^3.0.1",
-    "@typescript-eslint/parser": "^3.0.1",
+    "@typescript-eslint/eslint-plugin": "^3.1.0",
+    "@typescript-eslint/parser": "^3.1.0",
     "@zeit/next-typescript": "^1.1.1",
     "babel-eslint": "^10.1.0",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.2",
-    "enzyme-to-json": "^3.4.4",
-    "eslint": "^7.0.0",
+    "enzyme-to-json": "^3.5.0",
+    "eslint": "^7.1.0",
     "eslint-config-airbnb": "^18.1.0",
     "eslint-config-airbnb-base": "^14.1.0",
     "eslint-plugin-babel": "^5.3.0",
     "eslint-plugin-import": "^2.20.2",
-    "eslint-plugin-jest": "^23.10.0",
+    "eslint-plugin-jest": "^23.13.2",
     "eslint-plugin-jsx-a11y": "^6.2.3",
     "eslint-plugin-prettier": "^3.1.3",
-    "eslint-plugin-react": "^7.19.0",
-    "eslint-plugin-react-hooks": "^4.0.0",
+    "eslint-plugin-react": "^7.20.0",
+    "eslint-plugin-react-hooks": "^4.0.4",
     "jest": "^26.0.1",
     "jest-mock-extended": "^1.0.8",
     "jsdom": "16.2.2",
     "jsdom-global": "3.0.2",
     "prettier": "^2.0.5",
-    "ts-jest": "^26.0.0",
-    "typescript": "^3.8.3"
+    "ts-jest": "^26.1.0",
+    "typescript": "^3.9.3"
   }
 }
