+++ pages/index/index.tsx
@@ -1,19 +1,7 @@
 import React, { ReactElement } from "react";
-import { Container } from "@material-ui/core";
 
-import { useRedirectIfLoggedOut } from "../../src/effects/auth";
-import ContainerView from "../../components/containerView";
+import IndexPage from "../../components/pages/index";
 
-export default function Home(): ReactElement {
-  useRedirectIfLoggedOut();
-
-  // TODO move Loading indicator into separate component
-  // TODO above todo was done but thinking we should
-  // move header and isloading.. spinner into one as well so that it's a different
-  // header if logged in etc...
-  return (
-    <Container>
-      <ContainerView />
-    </Container>
-  );
+export default function Index(): ReactElement {
+  return <IndexPage />;
 }
