+++ tsconfig.json
@@ -20,7 +20,10 @@
     "jsx": "preserve"
   },
   "exclude": [
-    "node_modules"
+    "node_modules",
+    "**/__mocks__/**",
+    "**/__snapshots__/**",
+    "lib/**"
   ],
   "include": [
     "next-env.d.ts",
